Transition vers le nouveau schema de BDD (transparent pour la partie Android) TODO: Si tout marche supprimer les anciennes tables 	modifié :         src/AppliEdition/app/src/main/java/BDD/db/DataBaseHelper.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/db/DataBaseManager.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/db/Synchronize.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/Alertes.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/Armature.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/Evenement.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/MesuresNonLues.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/Reprise.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/Suspension.java 	modifié :         src/AppliEdition/app/src/main/java/BDD/to/VarRythmes.java 	modifié :         src/AppliEdition/app/src/main/java/com/example/boris/emaestro/MainActivity.java 	modifié :         src/Maestro Box/MySQL/create.sql

+++ src/Maestro Box/MySQL/create.sql
@@ -20,4 +20,4 @@ CREATE TABLE Alerte (id_alerte INTEGER PRIMARY KEY, id_musique INTEGER, mesure_d
 CREATE TABLE VariationRythme (id_variation_rythme INTEGER PRIMARY KEY, id_musique INTEGER, mesure_debut INTEGER, temps_debut INTEGER, taux_de_variation FLOAT, passage_reprise INTEGER);
 CREATE TABLE Suspension (id_suspension INTEGER PRIMARY KEY, id_musique INTEGER, mesure INTEGER, temps INTEGER, duree INTEGER, passage_reprise INTEGER);
 CREATE TABLE Armature (id_armature INTEGER PRIMARY KEY, id_musique INTEGER, mesure_debut INTEGER, temps_debut INTEGER, alteration INTEGER, passage_reprise INTEGER);
-CREATE TABLE Evenement (id_evenement INTEGER PRIMARY KEY, id_musique INTEGER, flag INTEGER, mesure_debut INTEGER, arg2 INTEGER, passage_reprise INTEGER, arg3 INTEGER);
+CREATE TABLE Evenement (id_evenement INTEGER PRIMARY KEY, id_musique INTEGER, flag INTEGER, mesure_debut INTEGER, arg2 INTEGER, passage_reprise INTEGER, arg3 FLOAT);
